Add pytest configuration and refactor test fixtures for monitoring service
