+++ pytest.ini
@@ -1,0 +1,11 @@
+[pytest]
+addopts = -v --tb=short
+pythonpath = .
+testpaths = tests
+python_files = test_*.py
+python_classes = Test*
+python_functions = test_*
+
+# Set asyncio plugin configuration
+asyncio_mode = strict
+asyncio_default_fixture_loop_scope = function
